- Add default branch check before deleting virtual branch directly

+++ gitbutler-ui/src/lib/components/BranchLanePopupMenu.svelte
@@ -48,7 +48,11 @@
 			<ContextMenuItem
 				label="Delete"
 				on:click={async () => {
-					if (commits.length == 0 && branch.files?.length == 0) {
+					if (
+						branch.name.toLowerCase().includes('virtual branch') &&
+						commits.length == 0 &&
+						branch.files?.length == 0
+					) {
 						await branchController.deleteBranch(branch.id);
 					} else {
 						deleteBranchModal.show(branch);
